Help menu overhaul-isa. Dealing with parsing.
The help menu has been moved into a command when the `help` keyword is
used now. It makes sense for me to use this convention.

The parsing also now happens after the block is evaluated before ARGV
emptiness is checked for the help menu or default help menu to run. The
`remove_default_help_menu` now allows the default help menu to be
turned off if called in an application block.